requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation and Models
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and Web
requests==2.31.0
httpx==0.25.2

# Logging
loguru==0.7.2

# OpenAI API
openai==1.3.8

# Web Automation
selenium==4.16.0
webdriver-manager==4.0.1

# Image Processing
Pillow==10.1.0

# Data Processing
pyyaml==6.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
aiofiles==23.2.1
jinja2==3.1.2
psutil==5.9.8

# Telegram Bot
python-telegram-bot==20.7
aiohttp>=3.8.0

# DNS utilities
dnspython==2.4.2
